feat: add payment confirmation screens
