+++ hooks/useGetPaymentStatus.tsx
@@ -4,7 +4,6 @@ import { useSubscription } from "@apollo/client"
 
 import { OperationError } from "../lib/graphql/index.types.d"
 import { LN_INVOICE_PAYMENT_STATUS } from "../lib/graphql/subscription"
-import { ACTIONS, ACTIONTYPE } from "../pages/merchant/_reducer"
 
 interface Props {
   paymentRequest: string | undefined
