+++ package.json
@@ -12,13 +12,6 @@
     "check-code": "yarn prettier-check && yarn eslint-check && yarn tsc-check"
   },
   "dependencies": {
-    "@apollo/client": "^3.4.16",
-    "@galoymoney/client": "^0.1.39",
-    "@types/node": "^16.11.1",
-    "@types/react": "^17.0.30",
-    "@types/react-dom": "^17.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.5.0",
-    "@typescript-eslint/parser": "^5.5.0",
     "bech32": "^2.0.0",
     "bitcoinjs-lib": "5.0.0",
     "bolt11": "1.3.4",
@@ -41,9 +34,6 @@
     "react-timer-hook": "^3.0.5",
     "react-to-print": "^2.12.6",
     "subscriptions-transport-ws": "^0.11.0",
-    "ts-pnp": "1.2.0",
-    "typescript": "^4.4.4",
-    "url": "^0.11.0",
     "use-debounce": "^7.0.1"
   },
   "devDependencies": {
@@ -60,6 +50,7 @@
     "@typescript-eslint/eslint-plugin": "^5.5.0",
     "@typescript-eslint/parser": "^5.5.0",
     "eslint": "^8.3.0",
+    "eslint-config-next": "^12.2.3",
     "eslint-config-react-app": "^6.0.0",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-import": "^2.26.0",
@@ -69,8 +60,8 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-testing-library": "^5.0.1",
     "eslint-webpack-plugin": "^3.0.1",
-    "react-dev-utils": "^12.0.1",
     "prettier": "^2.4.1",
+    "react-dev-utils": "^12.0.1",
     "ts-pnp": "1.2.0",
     "typescript": "^4.4.4"
   },
